V1.0
Enabled the ability to utilise a key that the user defines. Haven't
extensively bug-tested but it's a small program and it works.

Only note to self: Must find an easier way to clear the input stream
insteand of using std::cin.clear then std::cin.ignore(blahblah) every
time I need to flush it. Maybe put it in a function?

Anyway program is working, hard enough to break it and it does it's job
as well as simplicity allows.